issue002-chapter5-tictactoe-ui: Handle end of game notification and new game start
FEATURE(s)
issue002-chapter5-tictactoe-ui: Create a Jupyter notebook compatible
  tictactoe-ui

FILE(s)
[modified]src/nfn/samples/chapter5/tictactoe-ui/css/module.css:
  Define the styling for the overlay element in the tictactoe grid.
[modified]src/nfn/samples/chapter5/tictactoe-ui/index.html:
  Add the overlay element to the HTML structure.
[modified]src/nfn/samples/chapter5/tictactoe-ui/js/tictactoe-ui.mjs:
  Implement the logic to display the overlay with the game result and
  handle new game start.

DESCRIPTION
Handle end of game notification and new game start.

This creates displays an overlay indicating the end of the game and
the result (win/lose/draw). It also allows the user to start a new
game by clicking anywhere on the overlay.

Still a little unpolished with respect to the UI and the code is
not very well structured, but it works.